Moved xmtp files to dedicated directory.
