+++ src/App.tsx
@@ -26,7 +26,7 @@ import Messaging from './pages/Messaging';
 import { PushProvider } from './messaging/push/context/pushUser';
 import { chain, configureChains, createClient, WagmiConfig } from 'wagmi';
 import Messaging from './pages/Messaging';
-import { XmtpContextProvider } from './messaging/context/XmtpContext';
+import { XmtpContextProvider } from './messaging/xmtp/context/XmtpContext';
 
 const chains: Chain[] = [customChains.polygonMumbai];
